Restructure folder hierarchy, uses lowercase naming convention for folders
