+++ client/we_meet_client/src/components/Views/MainView.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
-import Hero from "../ui-components/Hero";
-import EventsCarousel from "../ui-components/EventsCarousel";
-import ExploreCategories from "../ui-components/ExploreCategories";
+import Hero from "../ui-components/main-view/Hero";
+import EventsCarousel from "../ui-components/main-view/EventsCarousel";
+import ExploreCategories from "../ui-components/main-view/ExploreCategories";
 
 class MainView extends Component {
   render() {
