+++ client/we_meet_client/src/components/ui-components/main-view/EventsCarousel.js
